Makefile.inc: Use pkg-config to find libraries
MacOS systems using MacPorts, among others, do not have libpng and
liblz4 include files and libraries in the standard system paths; this
ensures that wherever they are, they're added to the compiler and
linker options. There are, however, dozens of Makefiles in subdirs
that override the CFLAGS/LFLAGS settings in the top-level Makefile.inc;
for the ones that need the pkg-config settings, we change the `CFLAGS =`
to `CFLAGS += ` with the additional options they add beyond what
Makefile.inc originally had.

This requires that users have installed `pkg-config`, but most
systems with development tools will have that already.

Also add a `show-config` target that lets us easily check the generated
values for CFLAGS and LFLAGS.